Ref: 現在地取得providerをcommonに移動
