+++ lib/common/const/location.dart
@@ -1,4 +1,4 @@
 class LocationConstant {
-  static double tokyoStationLatitude = 35.681236;
-  static double tokyoStationLongitude = 139.767125;
+  static const double tokyoStationLatitude = 35.681236;
+  static const double tokyoStationLongitude = 139.767125;
 }
