--- lib/features/location/model/location_model.dart
@@ -1,16 +1,0 @@
-
-import 'package:freezed_annotation/freezed_annotation.dart';
-
-part 'location_model.freezed.dart';
-part 'location_model.g.dart';
-
-@freezed
-class LocationModel with _$LocationModel{
-  factory LocationModel({
-    @Default(0) double latitude,
-    @Default(0) double longitude
-  }) = _LocationModel;
-
-  factory LocationModel.fromJson(Map<String, dynamic> json) =>
-      _$LocationModelFromJson(json);
-}
