--- lib/pages/google_map_page/widget/charger_card_list_parts/charger_card_image.dart
@@ -1,21 +1,0 @@
-import 'package:cached_network_image/cached_network_image.dart';
-import 'package:flutter/material.dart';
-
-import '../../../../gen/assets.gen.dart';
-
-class ChargerCardImage extends StatelessWidget {
-  const ChargerCardImage({super.key, required this.images});
-
-  final List<String> images;
-
-  @override
-  Widget build(BuildContext context) {
-    return images.isNotEmpty
-        ? CachedNetworkImage(
-            width: double.infinity,
-            imageUrl: images[0],
-            fit: BoxFit.fitWidth,
-          )
-        : Assets.images.noImage.image(fit: BoxFit.fill);
-  }
-}
